pyproject.toml
[project]
name = "study-htmx-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.127.0",
    "openai-agents>=0.6.4",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
